Zelda Game
With a very raw multiplayer feature 
Note : The multiplayer is very raw and has only visual features implemented, making a full-fledged system takes a lot of time, and time that I as a 10th grader don't have. 
I have referred to this while coding. Assets are available here as well. Note : Some assets may be different forexample I have a Frog system implemented with a few other graphics and audio which may not be included in the video.